Fix for display of choices larger than the current window.
It seems that the initial window height was not being taken in account
during init and the choices were set to all choices and they overflowed
giving a seamingly bad list until a filter was applied or the window
was resized.

This calls `m.forceUpdatePageSizeForHeight()` on the model `Init()`
which inits `m.currentChoices` and `m.availableChoices` correctly.

Fixes: #18